+++ index.html
@@ -226,9 +226,6 @@
       </div>
     </section>
   </template>
-  <script src="js/function.js">
-  </script>
-  <script src="js/main.js">
-  </script>
+  <script src="js/main.js" type="module"></script>
 </body>
 </html>
