+++ js/data.js
@@ -1,0 +1,40 @@
+// Массив с описаниями
+const DESCRIPTION = [
+  'Как-то так!',
+  'Какая прелесть!',
+  'Захотелось запечатлить эту картину',
+  'Я думаю этот кадр прекрасен!',
+  'Как вам такое?',
+  'Вот это я поймал кадр!!!',
+  'Хейтеры - мимо =)'
+];
+
+//Массив с комментариями
+const MESSAGE_TEXT = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+// Массив имен комментаторов
+const NAMES = [
+  'Александр',
+  'Никита',
+  'Андрей',
+  'Виктор',
+  'Анна',
+  'Ольга'
+];
+
+//Константы
+const NUMBER_PHOTOS_ON_PAGE = 25;
+const MAX_NUMBER_COMMENTS = 30;
+const MIN_NUMBER_LIKES = 15;
+const MAX_NUMBER_LIKES = 200;
+const MIN_NUMBER_MESSAGES = 1;
+const MAX_NUMBER_MESSAGES = 2;
+
+export {DESCRIPTION, MESSAGE_TEXT, NAMES, NUMBER_PHOTOS_ON_PAGE, MAX_NUMBER_COMMENTS, MAX_NUMBER_LIKES, MAX_NUMBER_MESSAGES, MIN_NUMBER_LIKES, MIN_NUMBER_MESSAGES};
